+++ src/app/components/ProductionCard/index.tsx
